requirements.txt
Authlib==1.6.0
asyncpg==0.30.0
fastapi==0.116.1
itsdangerous==2.2.0
python-jose==3.5.0
psycopg2-binary==2.9.10
pydantic==1.10.22
python-dotenv==1.1.1
SQLAlchemy==2.0.41
uvicorn==0.35.0